Improve documentation and clarity
